Rename SelectableColor to ColorSelector and add Component after component names
