+++ src/Components/BottomBarComponent/Components/ColorSelector/ColorSelector.css
