+++ src/Components/CanvasComponent/Components/PixelInfo.css
