+++ src/Components/CanvasComponent/Components/PixelInfo.js
