+++ src/Components/CanvasComponent/canvasUtils.js
