+++ src/Components/KickDialogComponent/KickDialog.js
